feat: types/mydashboard 생성 후 적용 및  type import 변경
