+++ src/components/dashboard/dashboard-button/index.tsx
@@ -1,10 +1,10 @@
 import Image from 'next/image'
 import { useRouter } from 'next/router'
 
-import { DashboardType } from '@/pages/mydashboard'
 import addBoardBtn from '@/public/icons/add-board-btn.svg'
 import crownIcon from '@/public/icons/crown-icon.svg'
 import rightArrow from '@/public/icons/right-arrow.svg'
+import { DashboardType } from '@/src/types/mydashboard'
 
 import S from './dashboardButton.module.scss'
