+++ src/types/mydashboard.ts
@@ -1,0 +1,9 @@
+export interface DashboardType {
+  id: number
+  title: string
+  color: string
+  userId: number
+  createdAt: string
+  updatedAt: string
+  createdByMe: true
+}
